API refactor

+++ api/views.py
@@ -1,6 +1,4 @@
-from django.http import Http404
-
-from rest_framework import authentication, permissions
+from rest_framework import authentication, permissions, status
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework.views import APIView
@@ -23,7 +21,7 @@ class mealList(APIView):
         meals = Meal.objects.filter(user=request.user)
         serialiser = MealSerializer(meals, many=True)
         return Response({
-            'status': 'OK',
+            'status': status.HTTP_200_OK,
             'items': serialiser.data,
         })
 
@@ -35,9 +33,14 @@ class mealList(APIView):
         if serialiser.is_valid():
             serialiser.save()
         else:
-            return Response(serialiser.errors)
+            return Response({
+                serialiser.errors
+            })
 
-        return Response(serialiser.data)
+        return Response({
+            'status': status.HTTP_200_OK,
+            'item': serialiser.data,
+        })
 
 
 class mealDetail(APIView):
@@ -45,37 +48,65 @@ class mealDetail(APIView):
     permission_classes = [permissions.IsAuthenticated]
 
     def get_object(self, pk):
+        #print(rest_framework.status.HTTP_404_NOT_FOUND)
         try:
             return Meal.objects.get(id=pk)
         except Meal.DoesNotExist:
-            raise Http404
+            return None
 
     def delete(self, request, pk, format=None):
         meal = self.get_object(pk=pk)
 
+        if meal is None:
+            return ResponseApiError(status.HTTP_404_NOT_FOUND)
+
+        if meal.user.id is not request.user.id:
+            return ResponseApiError()
+
         if request.method == 'DELETE':
             meal.delete()
             return Response({
-                "ok": "ok"
+                "status": status.HTTP_200_OK
             })
 
     def get(self, request, pk, format=None):
         meal = self.get_object(pk=pk)
+
+        if meal is None:
+            return ResponseApiError(status.HTTP_404_NOT_FOUND)
+
+        if meal.user.id is not request.user.id:
+            return ResponseApiError()
         
         serialiser = MealSerializer(meal, many=False)
-        return Response(serialiser.data)
+        return Response({
+            "status": status.HTTP_200_OK, 
+            "item": serialiser.data
+        })
 
     def post(self, request, pk, format=None):
         meal = self.get_object(pk=pk)
 
-        serializer = MealSerializer(instance=meal, data=request.data)
+        if meal is None:
+            return ResponseApiError(status.HTTP_404_NOT_FOUND)
+
+        if meal.user.id is not request.user.id:
+            return ResponseApiError()
+
+        data = request.data
+        data['user'] = request.user.id
+
+        serializer = MealSerializer(instance=meal, data=data)
 
         if serializer.is_valid():
             serializer.save()
         else:
-            return Response(serialiser.errors)
+            return ResponseApiSerializerError(serializer)
 
-        return Response(serializer.data)
+        return Response({
+            "status": status.HTTP_200_OK,
+            "item": serializer.data,
+        })
 
 
 class mealTypeList(APIView):
@@ -96,7 +127,7 @@ class mealTypeList(APIView):
         if serialiser.is_valid():
             serialiser.save()
         else:
-            return Response(serialiser.errors)
+            return ResponseApiSerializerError(serializer)
 
         return Response(serialiser.data)
 
@@ -122,7 +153,7 @@ class mealTypeDetail(APIView):
         if request.method == 'DELETE':
             type.delete()
             return Response({
-                "ok": "ok"
+                status: status.HTTP_200_OK
             })
 
     def post(self, request, pk, format=None):
@@ -136,3 +167,31 @@ class mealTypeDetail(APIView):
             return Response(serialiser.errors)
 
         return Response(serializer.data)
+
+
+def ResponseApiError(status_code=status.HTTP_401_UNAUTHORIZED):
+    message = {
+        "status": status_code,
+        "title": "Unauthorized",
+        "message": "The request has not been applied because it lacks valid authentication credentials for the target resource."
+    }
+
+    if status_code == status.HTTP_404_NOT_FOUND:
+        message['title'] = "Not Found"
+        message['message'] = "The request has not been found."
+
+    return Response(message, status=status_code)
+
+
+def ResponseApiSerializerError(serializer, status_code=status.HTTP_400_BAD_REQUEST):
+    message = {
+        "status": status_code,
+        "title": "Bad Request",
+        "message": ""
+    }
+
+    for field in serializer.errors:
+        for error in serializer.errors[field]:
+            message['message'] += field + " : " + error + "\n"
+
+    return Response(message, status=status_code)
